requirements.txt
moviepy
numpy
pytesseract
librosa
opencv-python
Pillow
soundfile
pytest
